// src/models/interface/Labour.interface.ts
export default interface labourAttributes {
    id: number;
    supervisorId: number;
    userProfile: string;
    firstName: string;
    middleName: string;
    lastName: string;
    nameAsPerAadharCard: string;
    gender: string
    dob: Date;
    maritalStatus: string;
    fatherName: string;
    husbandName: string;
    mobileNumber: string;
    email: string;
    knownLanguage: string;
    religion: string;
    occupation: string;
    natureOfWork: string
    category: string;
    department: string;
    dateOfLeave: Date;
    emergencyMobileNumber: string;
    nomineeDetails: string;
    presentAddress: string;
    permanentAddress: string;
    providentFundNumber: string;
    universalAccountNumber: string;
    employeeStateInsurance: string;
    bloodGroup: string;
    previousOrganizationName: string;
    dateOfExitPreviousEmployer: Date;
    reasonOfLeave: string;
    workExperience: string;
    covidVaccine: boolean;
    aadharNumber: string;
    panCardNumber: string;
    nearestPoliceStation: string;
    cardNumber: string;
    bankAccountNumber: string;
    accountHolderName: string;
    bankName: string;
    ifcCode: string;
    branchName: string;
    hireDate: Date;
    reference: string;
    relationshipWithReference: string;
    referenceContactNumber: string;
}